Move generators to utils.py
