+++ code/utils.py
@@ -1,3 +1,5 @@
+import copy
+import glob
 import json
 import os
 
@@ -37,3 +39,30 @@ def read_json_file(filename: str, empty_if_missing=False) -> dict:
         return {}
     with open(filename, encoding="utf-8") as file:
         return json.load(file)
+
+
+def get_all_geojson_files(show_progress: bool = True, rewrite_geojson: bool = False):
+    """A generator that returns (filename, geojson_data) for each GeoJSON file in the results directory"""
+    filenames = glob.glob("results/V**/s*.geojson")  # FIXME
+    for n, filename in enumerate(filenames):
+        if show_progress and n % 100 == 0:
+            print_progress_bar(n, len(filenames), prefix="Progress:", suffix="Complete", length=50)
+        geojson_data = read_json_file(filename)
+        if rewrite_geojson:
+            # take a copy of the GeoJSON, and if it is modified, write it back to the original file
+            geojson_data_copy = copy.deepcopy(geojson_data)
+            yield filename, geojson_data
+            if geojson_data != geojson_data_copy:
+                write_json_file(filename, geojson_data, indent=1)
+        else:
+            yield filename, geojson_data
+
+    # final 100% output
+    print_progress_bar(1, 1, prefix="Progress:", suffix="Complete", length=50)
+
+
+def get_all_features(show_progress: bool = True, rewrite_geojson: bool = False):
+    """A generator that returns (filename, geojson_data, feature) for every Feature in every GeoJSON file."""
+    for filename, geojson_data in get_all_geojson_files(show_progress, rewrite_geojson):
+        for feature in geojson_data["features"]:
+            yield filename, geojson_data, feature
